+++ src/Symfony/Component/Form/Tests/AbstractBootstrap3LayoutTest.php
@@ -118,12 +118,12 @@ abstract class AbstractBootstrap3LayoutTest extends AbstractLayoutTest
             [@class="list-unstyled"]
             [
                 ./li
-                    [.="[trans]Error 1[/trans]"]
+                    [.=" [trans]Error 1[/trans]"]
                     [
                         ./span[@class="glyphicon glyphicon-exclamation-sign"]
                     ]
                 /following-sibling::li
-                    [.="[trans]Error 2[/trans]"]
+                    [.=" [trans]Error 2[/trans]"]
                     [
                         ./span[@class="glyphicon glyphicon-exclamation-sign"]
                     ]
@@ -1334,7 +1334,10 @@ abstract class AbstractBootstrap3LayoutTest extends AbstractLayoutTest
         );
     }
 
-    public function testReadOnly()
+    /**
+     * @group legacy
+     */
+    public function testLegacyReadOnly()
     {
         $form = $this->factory->createNamed('name', 'text', null, array(
             'read_only' => true,
@@ -1899,14 +1902,13 @@ abstract class AbstractBootstrap3LayoutTest extends AbstractLayoutTest
         $form = $this->factory->createNamed('text', 'text', 'value', array(
             'required' => true,
             'disabled' => true,
-            'read_only' => true,
-            'attr' => array('maxlength' => 10, 'pattern' => '\d+', 'class' => 'foobar', 'data-foo' => 'bar'),
+            'attr' => array('readonly' => true, 'maxlength' => 10, 'pattern' => '\d+', 'class' => 'foobar', 'data-foo' => 'bar'),
         ));
 
         $html = $this->renderWidget($form->createView());
 
         // compare plain HTML to check the whitespace
-        $this->assertSame('<input type="text" id="text" name="text" readonly="readonly" disabled="disabled" required="required" maxlength="10" pattern="\d+" class="foobar form-control" data-foo="bar" value="value" />', $html);
+        $this->assertSame('<input type="text" id="text" name="text" disabled="disabled" required="required" readonly="readonly" maxlength="10" pattern="\d+" class="foobar form-control" data-foo="bar" value="value" />', $html);
     }
 
     public function testWidgetAttributeNameRepeatedIfTrue()