+++ src/Symfony/Component/Translation/Loader/JsonFileLoader.php
@@ -25,10 +25,21 @@ class JsonFileLoader extends FileLoader
      */
     protected function loadResource($resource)
     {
-        $messages = json_decode(file_get_contents($resource), true);
+        if (!stream_is_local($resource)) {
+            throw new InvalidResourceException(sprintf('This is not a local file "%s".', $resource));
+        }
 
-        if (0 < $errorCode = json_last_error()) {
-            throw new InvalidResourceException(sprintf('Error parsing JSON - %s', $this->getJSONErrorMessage($errorCode)));
+        if (!file_exists($resource)) {
+            throw new NotFoundResourceException(sprintf('File "%s" not found.', $resource));
+        }
+
+        $messages = array();
+        if ($data = file_get_contents($resource)) {
+            $messages = json_decode($data, true);
+
+            if (0 < $errorCode = json_last_error()) {
+                throw new InvalidResourceException(sprintf('Error parsing JSON - %s', $this->getJSONErrorMessage($errorCode)));
+            }
         }
 
         return $messages;